Remove loading config from Env Variables
